refactor: 모두의 정원 VO 래핑
