+++ backend/pium/src/main/java/com/official/pium/domain/vo/WaterCycle.java
@@ -1,4 +1,4 @@
-package com.official.pium.domain;
+package com.official.pium.domain.vo;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
