+++ backend/pium/src/main/java/com/official/pium/service/dto/GardenCreateRequest.java
@@ -1,8 +1,12 @@
 package com.official.pium.service.dto;
 
+import com.official.pium.domain.Garden;
+import com.official.pium.domain.PetPlant;
+import com.official.pium.domain.vo.PlantState;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
+import java.time.LocalDate;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -24,4 +28,25 @@ public class GardenCreateRequest {
 
     @NotBlank
     private String manageLevel;
+
+    public Garden toGarden(PetPlant petPlant) {
+        return Garden.builder()
+                .dictionaryPlant(petPlant.getDictionaryPlant())
+                .member(petPlant.getMember())
+                .nickname(petPlant.getNickname())
+                .imageUrl(petPlant.getImageUrl())
+                .plantState(
+                        PlantState.builder()
+                                .location(petPlant.getLocation())
+                                .flowerpot(petPlant.getFlowerpot())
+                                .light(petPlant.getLight())
+                                .wind(petPlant.getWind())
+                                .build()
+                )
+                .daySince(petPlant.calculateDaySince(LocalDate.now()))
+                .waterCycle(petPlant.getWaterCycle())
+                .content(this.content)
+                .manageLevel(this.manageLevel)
+                .build();
+    }
 }
