+++ backend/pium/src/main/java/com/official/pium/service/dto/SingleGardenResponse.java
@@ -1,5 +1,6 @@
 package com.official.pium.service.dto;
 
+import com.official.pium.domain.Garden;
 import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -21,6 +22,32 @@ public class SingleGardenResponse {
     private String manageLevel;
     private PetPlantInfo petPlant;
 
+    public static SingleGardenResponse from(Garden garden) {
+        return SingleGardenResponse.builder()
+                .id(garden.getId())
+                .createdAt(garden.getCreatedAt())
+                .updatedAt(garden.getUpdatedAt())
+                .dictionaryPlantName(garden.getDictionaryPlant().getClassification()
+                        .getName())
+                .content(garden.getContent())
+                .manageLevel(garden.getManageLevel())
+                .petPlant(toPetPlantInfo(garden))
+                .build();
+    }
+
+    private static SingleGardenResponse.PetPlantInfo toPetPlantInfo(Garden garden) {
+        return SingleGardenResponse.PetPlantInfo.builder()
+                .imageUrl(garden.getImageUrl())
+                .nickname(garden.getNickname())
+                .location(garden.getPlantState().getLocation())
+                .flowerpot(garden.getPlantState().getFlowerpot())
+                .light(garden.getPlantState().getLight())
+                .wind(garden.getPlantState().getWind())
+                .daySince(garden.getDaySince())
+                .waterCycle(garden.getWaterCycle())
+                .build();
+    }
+
     @Getter
     @NoArgsConstructor
     public static class PetPlantInfo {
@@ -35,7 +62,8 @@ public class SingleGardenResponse {
         private Integer waterCycle;
 
         @Builder
-        private PetPlantInfo(String imageUrl, String nickname, String location, String flowerpot, String light, String wind, Long daySince, Integer waterCycle) {
+        private PetPlantInfo(String imageUrl, String nickname, String location, String flowerpot, String light,
+                             String wind, Long daySince, Integer waterCycle) {
             this.imageUrl = imageUrl;
             this.nickname = nickname;
             this.location = location;
